requirements.txt
faiss-cpu
numpy
sentence-transformers
langchain-text-splitters
google-genai
requests
python-dotenv
pymupdf
python-docx
python-pptx
openpyxl
einops
